Update RV Masses Code.py
I added the varying parameters and created a list of parameters to vary. I am thinking that we can use that list to have the user decide which parameters they would want to vary. I also added the telescope as an attribute in the Target class so that it can be used the posterior parameters in the radvel fitting portion.